chore(cli): moved package under "webpack-cli" folder
