+++ packages/webpack-cli/__tests__/command/external.test.js
