+++ packages/webpack-cli/lib/utils/Compiler.js
