+++ packages/webpack-cli/lib/utils/CompilerOutput.js
