+++ packages/webpack-cli/lib/utils/GroupHelper.js
