+++ packages/webpack-cli/lib/utils/production-config.js
