+++ test/utils/test-utils.js
@@ -4,7 +4,7 @@ const fs = require('fs');
 const execa = require('execa');
 const { sync: spawnSync } = execa;
 const { Writable } = require('readable-stream');
-const WEBPACK_PATH = path.resolve(__dirname, '../../packages/cli/bin/cli.js');
+const WEBPACK_PATH = path.resolve(__dirname, '../../packages/webpack-cli/bin/cli.js');
 const ENABLE_LOG_COMPILATION = process.env.ENABLE_PIPE || false;
 /**
  * Description
